Disabled Energy Ghosts, tamed FlimFlam
Energy Ghosts have a slim chance of picking up armor, which will
disappear with them when dawn rolls around.  FlimFlam is fun, but some
of the features can be cruel if someone doesn't know what is going on.
As such, no ghosts, and adjusted flimflam to attempt to be kinder to
players.